+++ castle/cms/tests/test_metadata.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
-from castle.cms.behaviors.location import ILocation
-from castle.cms.tiles.metadata import MetaDataTile
-from castle.cms.testing import CASTLE_PLONE_INTEGRATION_TESTING
-from plone import api
-from plone.app.testing import login
-from plone.app.testing import setRoles
-from plone.app.testing import TEST_USER_ID
-from plone.app.testing import TEST_USER_NAME
-
 import unittest
+
+from castle.cms.behaviors.location import ILocation
+from castle.cms.behaviors.search import ISearch
+from castle.cms.testing import CASTLE_PLONE_INTEGRATION_TESTING
+from castle.cms.tiles.metadata import MetaDataTile
+from plone import api
+from plone.app.testing import TEST_USER_ID, TEST_USER_NAME, login, setRoles
 
 
 class TestMetadata(unittest.TestCase):
@@ -31,3 +29,41 @@ class TestMetadata(unittest.TestCase):
         result = tile()
         self.assertTrue('Green Bay, WI' in result)
         self.assertTrue('name="location"' in result)
+
+    def test_robots_default_config(self):
+        page = api.content.create(type='Document', title='Foobar',
+                                  container=self.portal)
+
+        tile = MetaDataTile(page, self.request)
+        result = tile()
+        self.assertTrue('index,follow' in result)
+
+    def test_robots_no_value(self):
+        page = api.content.create(type='Document', title='Foobar',
+                                  container=self.portal)
+
+        search = ISearch(page)
+        search.robot_configuration = [u'follow']
+
+        tile = MetaDataTile(page, self.request)
+        result = tile()
+        self.assertTrue('follow,noindex' in result)
+
+        search = ISearch(page)
+        search.robot_configuration = [u'index']
+
+        tile = MetaDataTile(page, self.request)
+        result = tile()
+        self.assertTrue('index,nofollow' in result)
+
+    def test_robots_values(self):
+        page = api.content.create(type='Document', title='Foobar',
+                                  container=self.portal)
+
+        search = ISearch(page)
+        search.robot_configuration = [
+            u'index', u'follow', u'noimageindex', u'noarchive', u'nosnippet']
+
+        tile = MetaDataTile(page, self.request)
+        result = tile()
+        self.assertTrue('index,follow,noimageindex,noarchive,nosnippet' in result)